+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/HDBakedReflectionSystem.cs
@@ -635,7 +635,7 @@ namespace UnityEditor.Rendering.HighDefinition
                         importer.sRGBTexture = false;
                         importer.filterMode = FilterMode.Bilinear;
                         importer.mipmapEnabled = false;
-                        importer.textureCompression = hd.currentPlatformRenderPipelineSettings.lightLoopSettings.planarReflectionCacheCompressed
+                        importer.textureCompression = false
                             ? TextureImporterCompression.Compressed
                             : TextureImporterCompression.Uncompressed;
                         importer.textureShape = TextureImporterShape.Texture2D;